backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
psycopg2-binary
sentence-transformers
faiss-cpu
pydantic>=2.0
cachetools
python-multipart
python-docx
pypdf
pyyaml
google-generativeai
python-dotenv
pandas
pytest
pytest-mock